+++ com.unity.render-pipelines.high-definition/Editor/DefaultScene/HDSceneManagement.cs
@@ -39,8 +39,14 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
 
         static void NewSceneCreated(Scene scene, NewSceneSetup setup, NewSceneMode mode)
         {
-            if (!InHDRP() || HDProjectSettings.defaultScenePrefab == null)
+            if (!InHDRP())
                 return; // do not interfere outside of hdrp
+
+            if (HDProjectSettings.defaultScenePrefab == null)
+            {
+                Debug.LogError("Default Scene not set! Please run Wizard...");
+                return;
+            }
 
             if (setup == NewSceneSetup.DefaultGameObjects)
             {
@@ -109,6 +115,12 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
         {
             public override void Action(int instanceId, string pathName, string resourceFile)
             {
+                if (HDProjectSettings.defaultScenePrefab == null)
+                {
+                    Debug.LogError("Default Scene not set! Please run Wizard...");
+                    return;
+                }
+
                 if (s_CreateEmptySceneAsset(pathName))
                 {
                     UnityEngine.Object sceneAsset = AssetDatabase.LoadAssetAtPath(pathName, typeof(SceneAsset));